+++ CHANGELOG.md
@@ -23,6 +23,7 @@ v1.3.0-dev
  * ons-navigator: Added event handler attributes.
  * ons-popover: Added event handler attributes.
  * ons-pull-hook: Added event handler attributes.
+ * ons-split-view: Added event handler attributes.
 
 v1.2.2
 ----